fix(#92): 여행지역 추가 팝업 조건 변경
