+++ src/recoil/mytrip/selectedPlacesState.ts
@@ -4,7 +4,7 @@ interface Location {
     thumbnail?: string,
     id:number,
     name:string,
-    location: string[],
+    location: string,
 }
 
 export const selectedPlacesState = atom<Location[]>({
